Separate out Cassandra span matcher
